# docs/v1/swagger.yaml
basePath: /v1
definitions:
...
  request.Avatar:
    properties:
      key:
        description: 前端透传给后端用于保存avatar，例如：custom-upload/avatar/abc/def.png
        type: string
      path:
        description: 前端请求地址，例如：/v1/static/avatar/abc/def.png (请求非必填)
        type: string
    type: object
...
  request.Login:
    properties:
      code:
        description: 验证码
        type: string
      key:
        description: 客户端key
        type: string
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    required:
    - code
    - key
    - password
    - username
    type: object
...
  request.RegisterByEmail:
    properties:
      code:
        description: 邮箱验证码
        type: string
      email:
        description: 邮箱
        type: string
      username:
        description: 用户名
        type: string
    required:
    - code
    - email
    - username
    type: object
  request.RegisterSendEmailCode:
    properties:
      email:
        description: 邮箱
        type: string
      username:
        description: 用户名
        type: string
    required:
    - email
    - username
    type: object
  request.ResetPasswordByEmail:
    properties:
      code:
        description: 邮箱验证码
        type: string
      email:
        description: 邮箱
        type: string
      password:
        description: 密码
        type: string
    required:
    - code
    - email
    - password
    type: object
  request.ResetPasswordSendEmailCode:
    properties:
      email:
        description: 邮箱
        type: string
    required:
    - email
    type: object
...
  response.Captcha:
    properties:
      b64:
        description: 验证码png图片base64字符串
        type: string
      key:
        description: 客户端key
        type: string
    type: object
...
  response.ClientOverView:
    properties:
      activeClient:
        allOf:
        - $ref: '#/definitions/response.ClientOverviewItem'
        description: 活跃客户端
      additionClient:
        allOf:
        - $ref: '#/definitions/response.ClientOverviewItem'
        description: 新增客户端
    type: object
  response.ClientOverviewItem:
    properties:
      periodOverPeriod:
        description: 环比上周期百分比
        type: number
      value:
        description: 数量
        type: number
    type: object
  response.ClientStatistic:
    properties:
      overview:
        allOf:
        - $ref: '#/definitions/response.ClientOverView'
        description: 客户端统计面板
      trend:
        allOf:
        - $ref: '#/definitions/response.ClientTrends'
        description: 客户端统计趋势
    type: object
  response.ClientTrends:
    properties:
      client:
        allOf:
        - $ref: '#/definitions/response.StatisticChart'
        description: 客户端活跃数据
    type: object
...
  response.CustomAbout:
    properties:
      copyright:
        description: 版权
        type: string
      logoPath:
        description: 关于图标路径
        type: string
      version:
        type: string
    type: object
  response.CustomDefaultIcon:
    properties:
      agentIcon:
        type: string
      ragIcon:
        type: string
      workflowIcon:
        type: string
    type: object
  response.CustomEmail:
    properties:
      status:
        type: boolean
    type: object
  response.CustomHome:
    properties:
      backgroundColor:
        description: 平台背景色
        type: string
      logo:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 首页logo
      title:
        description: 平台名称
        type: string
    type: object
  response.CustomLogin:
    properties:
      background:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 登录页背景图
      loginButtonColor:
        description: 登录按钮颜色
        type: string
      logo:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 登录页图标
      platformDesc:
        description: 平台描述词
        type: string
      welcomeText:
        description: 登录页欢迎标词
        type: string
    type: object
  response.CustomLoginEmail:
    properties:
      email:
        allOf:
        - $ref: '#/definitions/response.CustomEmail'
        description: 登录邮箱
    type: object
  response.CustomRegister:
    properties:
      email:
        allOf:
        - $ref: '#/definitions/response.CustomEmail'
        description: 注册邮箱
    type: object
  response.CustomResetPassword:
    properties:
      email:
        allOf:
        - $ref: '#/definitions/response.CustomEmail'
        description: 邮箱
    type: object
  response.CustomTab:
    properties:
      logo:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 标签页图标
      title:
        description: 标签页标题
        type: string
    type: object
...
  response.GetWorkflowTemplateListResp:
    properties:
      downloadLink:
        $ref: '#/definitions/response.WorkflowTemplateURL'
      list:
        items:
          $ref: '#/definitions/response.WorkflowTemplateInfo'
        type: array
      total:
        type: integer
    type: object
  response.IDName:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  response.Language:
    properties:
      code:
        description: 语言代码
        type: string
      name:
        description: 语言名称
        type: string
    type: object
...
  response.ListResult:
    properties:
      list: {}
      total:
        type: integer
    type: object
  response.Login:
    properties:
      expireIn:
        type: string
      expiresAt:
        type: integer
      isUpdatePassword:
        description: 是否已更新密码
        type: boolean
      language:
        allOf:
        - $ref: '#/definitions/response.Language'
        description: 语言
      nickname:
        type: string
      orgPermission:
        allOf:
        - $ref: '#/definitions/response.UserOrgPermission'
        description: 用户所在组织权限
      orgs:
        description: 用户所在组织列表
        items:
          $ref: '#/definitions/response.IDName'
        type: array
      token:
        type: string
      uid:
        type: string
      username:
        type: string
    type: object
  response.LoginByEmail:
    properties:
      isEmailCheck:
        type: boolean
      isUpdatePassword:
        description: 是否已更新密码
        type: boolean
      token:
        type: string
    type: object
  response.LogoCustomInfo:
    properties:
      about:
        allOf:
        - $ref: '#/definitions/response.CustomAbout'
        description: 关于信息
      defaultIcon:
        allOf:
        - $ref: '#/definitions/response.CustomDefaultIcon'
        description: 应用默认图片
      home:
        allOf:
        - $ref: '#/definitions/response.CustomHome'
        description: 首页标题信息
      linkList:
        additionalProperties:
          type: string
        description: 跳转链接列表,key为链接名称,value为URL
        type: object
      login:
        allOf:
        - $ref: '#/definitions/response.CustomLogin'
        description: 登录页标题信息
      loginEmail:
        allOf:
        - $ref: '#/definitions/response.CustomLoginEmail'
        description: 邮箱登录信息
      register:
        allOf:
        - $ref: '#/definitions/response.CustomRegister'
        description: 注册信息
      resetPassword:
        allOf:
        - $ref: '#/definitions/response.CustomResetPassword'
        description: 重置密码信息
      tab:
        allOf:
        - $ref: '#/definitions/response.CustomTab'
        description: 标签页信息
    type: object
...
  response.Permission:
    properties:
      name:
        description: 权限名（对应菜单名）
        type: string
      perm:
        description: 权限
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  response.StatisticChart:
    properties:
      lines:
        description: 统计表中线段集合
        items:
          $ref: '#/definitions/response.StatisticChartLine'
        type: array
      tableName:
        description: 统计表名字
        type: string
    type: object
  response.StatisticChartLine:
    properties:
      items:
        description: 线段横纵坐标值
        items:
          $ref: '#/definitions/response.StatisticChartLineItem'
        type: array
      lineName:
        description: 线段名字
        type: string
    type: object
  response.StatisticChartLineItem:
    properties:
      key:
        type: string
      value:
        type: number
    type: object
...
  response.UserOrgPermission:
    properties:
      isAdmin:
        description: 是否系统内置管理员
        type: boolean
      isSystem:
        description: 是否系统视角（此时org.id为空，org.name为"系统"）
        type: boolean
      org:
        allOf:
        - $ref: '#/definitions/response.IDName'
        description: 组织
      permissions:
        description: 权限列表
        items:
          $ref: '#/definitions/response.Permission'
        type: array
      roles:
        description: 角色列表
        items:
          $ref: '#/definitions/response.IDName'
        type: array
    type: object
...
  response.WorkflowStatistic:
    properties:
      overview:
        allOf:
        - $ref: '#/definitions/response.WorkflowTemplateOverView'
        description: 工作流模板统计面板
      trend:
        allOf:
        - $ref: '#/definitions/response.WorkflowTemplateTrends'
        description: 工作流模板统计趋势
    type: object
  response.WorkflowTemplateDetail:
    properties:
      author:
        description: 作者
        type: string
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      category:
        description: 模板分类
        type: string
      desc:
        description: 模板描述
        type: string
      downloadCount:
        description: 下载次数
        type: integer
      feature:
        description: 模板特性说明
        type: string
      name:
        description: 模板名称
        type: string
      note:
        description: 注意事项
        type: string
      scenario:
        description: 模板应用场景
        type: string
      summary:
        description: 模板介绍概览
        type: string
      templateId:
        description: 模板ID
        type: string
    type: object
  response.WorkflowTemplateInfo:
    properties:
      author:
        description: 作者
        type: string
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      category:
        description: 模板分类
        type: string
      desc:
        description: 模板描述
        type: string
      downloadCount:
        description: 下载次数
        type: integer
      name:
        description: 模板名称
        type: string
      templateId:
        description: 模板ID
        type: string
    type: object
  response.WorkflowTemplateOverView:
    properties:
      browse:
        allOf:
        - $ref: '#/definitions/response.WorkflowTemplateOverviewItem'
        description: 工作流模板浏览数据总览
    type: object
  response.WorkflowTemplateOverviewItem:
    properties:
      periodOverPeriod:
        description: 环比上周期百分比
        type: number
      value:
        description: 数量
        type: number
    type: object
  response.WorkflowTemplateTrends:
    properties:
      browse:
        allOf:
        - $ref: '#/definitions/response.StatisticChart'
        description: 工作流模板浏览数据趋势
    type: object
  response.WorkflowTemplateURL:
    properties:
      url:
        type: string
    type: object
info:
  contact: {}
  description: "## HTTP Header\r\n| Header        | 说明      |\r\n| ------------- |
    --------- |\r\n| Authorization | JWT token |\r\n| X-Language    | 语言Code  |\r\n|
    X-Org-Id      | 组织ID    |\r\n| X-Client-Id   | 客户端标识|\r\n\r\n## HTTP Status\r\n|
    HTTP Status             | 说明                   |\r\n| -----------------------
    | ---------------------- |\r\n| 200, StatusOK           | 请求返回成功           |\r\n|
    400, StatusBadRequest   | 请求返回失败，用于业务 |\r\n| 401, StatusUnauthorized | JWT认证失败
    \           |\r\n| 403, StatusForbidden    | 没有权限               |\r\n\r\n## 权限-菜单对应表\r\n|
    一级权限        | 二级权限  | 三级权限 | 一级菜单 | 二级菜单 | 三级菜单 |\r\n|-------------|-------|------|------|------|------|\r\n|
    guest       |       |      | 【访客】 |      |      |\r\n| common      |       |      |
    【通用】 |      |      |\r\n| permission  |       |      | 权限管理 |      |      |\r\n|
    permission  | user  |      | 权限管理 | 用户管理 |      |\r\n| permission  | org   |      |
    权限管理 | 组织管理 |      |\r\n| permission  | role  |      | 权限管理 | 角色管理 |      |\r\n\r\n##
    `/v1/user/permission`返回用例\r\n```json\r\n{\r\n  \"code\": 0,\r\n  \"data\": {\r\n
    \   \"orgPermission\": {\r\n      \"org\": {\"id\": \"test-org-id\", \"name\":
    \"test-org-name\"},\r\n      \"permissions\": [\r\n        {\"perm\": \"permission\"},\r\n
    \       {\"perm\": \"permission.user\"},\r\n        {\"perm\": \"permission.org\"},\r\n
...
    \"操作成功\"\r\n}\r\n```"
...
  version: v0.0.1
paths:
...
  /base/captcha:
    get:
      consumes:
      - application/json
      parameters:
      - description: 语言
        in: header
        name: X-Language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.Captcha'
              type: object
      summary: 获取验证码
      tags:
      - guest
  /base/custom:
    get:
      parameters:
      - description: 语言
        in: header
        name: X-Language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.LogoCustomInfo'
              type: object
      summary: 自定义logo和title
      tags:
      - guest
...
  /base/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语言
        in: header
        name: X-Language
        type: string
      - description: 用户名+密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.Login'
              type: object
      summary: 用户登录
      tags:
      - guest
  /base/login/email:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户名+密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.LoginByEmail'
              type: object
      summary: 用户邮箱双因子登录
      tags:
      - guest
  /base/password/email:
    post:
      consumes:
      - application/json
      parameters:
      - description: 重置密码信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ResetPasswordByEmail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 邮箱重置密码
      tags:
      - guest
  /base/password/email/code:
    post:
      consumes:
      - application/json
      parameters:
      - description: 邮箱地址
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ResetPasswordSendEmailCode'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 重置密码邮箱验证码发送
      tags:
      - guest
  /base/register/email:
    post:
      consumes:
      - application/json
      parameters:
      - description: 邮箱注册信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RegisterByEmail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户邮箱注册
      tags:
      - guest
  /base/register/email/code:
    post:
      consumes:
      - application/json
      parameters:
      - description: 邮箱地址
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RegisterSendEmailCode'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 邮箱注册验证码发送
      tags:
      - guest
  /client/statistic:
    get:
      consumes:
      - application/json
      description: 获取使用工作流模板用户统计
      parameters:
      - description: 开始时间（格式yyyy-mm-dd）
        in: query
        name: startDate
        required: true
        type: string
      - description: 结束时间（格式yyyy-mm-dd）
        in: query
        name: endDate
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.ClientStatistic'
              type: object
      security:
      - JWT: []
      summary: 获取使用工作流模板用户统计
      tags:
      - common
  /client/statistic/cumulative:
    get:
      consumes:
      - application/json
      description: 获取累计使用工作流模板用户统计
      parameters:
      - description: 时间戳
        in: query
        name: endAt
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.ClientStatistic'
              type: object
      security:
      - JWT: []
      summary: 获取累计使用工作流模板用户统计
      tags:
      - common
...
  /workflow/template/detail:
    get:
      consumes:
      - application/json
      description: 获取工作流模板详情
      parameters:
      - description: 模板ID
        in: query
        name: templateId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.WorkflowTemplateDetail'
              type: object
      security:
      - JWT: []
      summary: 获取工作流模板详情
      tags:
      - guest
  /workflow/template/download:
    get:
      consumes:
      - application/json
      description: 下载工作流模板
      parameters:
      - description: 模板ID
        in: query
        name: templateId
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 下载工作流模板
      tags:
      - guest
  /workflow/template/list:
    get:
      consumes:
      - application/json
      description: 获取工作流模板列表
      parameters:
      - description: 模板名称
        in: query
        name: name
        type: string
      - description: 模板分类
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.GetWorkflowTemplateListResp'
              type: object
      security:
      - JWT: []
      summary: 获取工作流模板列表
      tags:
      - guest
  /workflow/template/recommend:
    get:
      consumes:
      - application/json
      description: 获取工作流模板推荐
      parameters:
      - description: 模板ID
        in: query
        name: templateId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.WorkflowTemplateInfo'
                        type: array
                    type: object
              type: object
      security:
      - JWT: []
      summary: 获取工作流模板推荐
      tags:
      - guest
  /workflow/template/statistic:
    get:
      consumes:
      - application/json
      description: 获取工作流模板统计
      parameters:
      - description: 开始时间（格式yyyy-mm-dd）
        in: query
        name: startDate
        required: true
        type: string
      - description: 结束时间（格式yyyy-mm-dd）
        in: query
        name: endDate
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.WorkflowStatistic'
              type: object
      security:
      - JWT: []
      summary: 获取工作流模板统计
      tags:
      - common
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"